Add ability to proxy trackmania api to another url
This will make it possible to run tests against a local simulator/mock for the trackmania api, but additionally make it so users of the plugin can stand up there own proxy api server to serve preloaded and cached api response data from the MX api. This allow users not as dependent on the uptime and reliability of the MX api when doing RMC/S runs.